requirements.txt
# Dev environment
pip
autopep8

# App
flask
Flask==2.1.2
torch
openai-whisper